requirements.txt
# Основные зависимости
python-telegram-bot==20.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25

# Для работы с временными зонами и датами
pytz==2023.3

# Для работы с PostgreSQL (закомментировано по умолчанию)
# psycopg2-binary==2.9.9 